# docs/swagger.yaml
info:
  contact: {}
paths:
  /api/v1/specification/list:
    get:
      consumes:
      - application/json
      description: Retrieve necessary environment variables
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get connector specification details
      tags:
      - specification
  /table/{tableName}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Enter table name
        in: path
        name: tableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Retrieve all records from a table
      tags:
      - table
    post:
      consumes:
      - application/json
      parameters:
      - description: Enter table name
        in: path
        name: tableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create a records for a table
      tags:
      - table
  /table/{tableName}/{sysId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Enter table name
        in: path
        name: tableName
        required: true
        type: string
      - description: Enter table record sys id
        in: path
        name: sysId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a record from a table
      tags:
      - table
    get:
      consumes:
      - application/json
      parameters:
      - description: Enter table name
        in: path
        name: tableName
        required: true
        type: string
      - description: Enter table record sys id
        in: path
        name: sysId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Retrieve a single record from a table
      tags:
      - table
    put:
      consumes:
      - application/json
      parameters:
      - description: Enter table name
        in: path
        name: tableName
        required: true
        type: string
      - description: Enter table record sys id
        in: path
        name: sysId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Modify a record from a table
      tags:
      - table
swagger: "2.0"